\name{tree}
\alias{tree}
\docType{data}

\title{Phylogenetic tree of the Melanoma dataset}
\description{
  Phylogenetic tree of the Melanoma dataset}
\source{Gopalakrishnan, V. , Spencer, C. N. , Nezi, L. , Reuben, A. , ,rews, M. C. , Karpinets, T. V. , Prieto, P. A. , Vicente, D. , Hoffman, K. , Wei, S. C. , Cogdill, A. P. , Zhao, L. , Hudgens, C. W. , Hutchinson, D. S. , Manzo, T. , Petaccia de Macedo, M. , Cotechini, T. , Kumar, T. , Chen, W. S. , Reddy, S. M. , Szczepaniak Sloane, R. , Galloway-Pena, J. , Jiang, H. , Chen, P. L. , Shpall, E. J. , Rezvani, K. , Alousi, A. M. , Chemaly, R. F. , Shelburne, S. , Vence, L. M. , Okhuysen, P. C. , Jensen, V. B. , Swennes, A. G. , McAllister, F. , Marcelo Riquelme Sanchez, E. , Zhang, Y. , Le Chatelier, E. , Zitvogel, L. , Pons, N. , Austin-Breneman, J. L. , Haydu, L. E. , Burton, E. M. , Gardner, J. M. , Sirmans, E. , Hu, J. , Lazar, A. J. , Tsujikawa, T. , Diab, A. , Tawbi, H. , Glitza, I. C. , Hwu, W. J. , Patel, S. P. , Woodman, S. E. , Amaria, R. N. , Davies, M. A. , Gershenwald, J. E. , Hwu, P. , Lee, J. E. , Zhang, J. , Coussens, L. M. , Cooper, Z. A. , Futreal, P. A. , Daniel, C. R. , Ajami, N. J. , Petrosino, J. F. , Tetzlaff, M. T. , Sharma, P. , Allison, J. P. , Jenq, R. R. and Wargo, J. A. (2018) Gut microbiome modulates response to anti-PD-1 immunotherapy in Melanoma patients. Science, 359 (6371), 97-103.}
